update repository: delete unused structs
