+++ src/ITLab-Projects/pkg/repositories/counter/counter.go
@@ -3,4 +3,7 @@ package counter
 type Counter interface {
 	Count() int64
 	UpdateCount() (int64, error)
+	CountByFilter(
+		filter interface{},
+	) (int64, error)
 }
